requirements/prod.txt
# Libraries for production use
celery==4.2.2
gevent==1.4.0
flask==1.0.2
requests==2.21.0
sqlalchemy==1.3.1
flask_sqlalchemy==2.3.2
gunicorn==19.9.0
python-dotenv==0.10.1
